forcedPhotCcd: implicit iteration over tracts
We want to iterate over all tracts spanned by the data when no tract is
provided by the user (who, presumably, doesn't want to have to look up what
tracts are spanned). This patch determines the set of tracts each visit
touches and adds a data reference with those tracts for each of the visit
components.

Ported from HSC-715.